+++ main/src/main/java/com/proyecto/ColectivosMapping.java
@@ -2,7 +2,7 @@ package com.proyecto;
 
 public class ColectivosMapping {
     
-    public Colectivos mapColectivos(int LineaBD, int UnidadesBD, Long IngresoMensualBD, Long MantenimientoBD){
+    public Colectivos mapColectivos(int LineaBD, int UnidadesBD, int IngresoMensualBD, int MantenimientoBD){
         Colectivos colectivos = new Colectivos(LineaBD, UnidadesBD, IngresoMensualBD, MantenimientoBD);
         return colectivos;
     }
